rename init package and configuration
