+++ startup/metrics.go
@@ -1,4 +1,4 @@
-package init
+package startup
 
 import (
 	"context"
